spell PlayerID correctly chars for bools for the passable attributes.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@16045 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/game/NetHandler.cxx
@@ -806,7 +806,7 @@ bool NetListener::listen (  Address serverAddress, unsigned short port )
   addr.sin_port = htons(port);
 
   // open well known service port
-  listenSocket = socket(AF_INET, SOCK_STREAM, 0);
+  listenSocket = (int)socket(AF_INET, SOCK_STREAM, 0);
   if (listenSocket == -1)
     return false;
